requirements.txt
# Backend dependencies
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
alembic==1.11.1
python-dotenv==1.0.0
email-validator==2.2.0

# Frontend dependencies
streamlit==1.42.2
requests==2.31.0

# Document processing
pdfplumber==0.11.5

# AI integration
google-generativeai==0.3.0
